+++ ui/src/actions/panels.js
@@ -3,6 +3,7 @@ export const UPDATE_PANEL_DATA = 'UPDATE_PANEL_DATA';
 export const ADD_PANEL = 'ADD_PANEL';
 export const ADD_PANEL_SET = 'ADD_PANEL_SET';
 export const DELETE_PANEL = 'DELETE_PANEL';
+export const CLONE_PANEL = 'CLONE_PANEL';
 
 function movePanelInData(data) {
 	return {
@@ -35,6 +36,17 @@ function addPanel(data) {
 
 export function addNewPanel(data) {
 	return dispatch => dispatch(addPanel(data));
+}
+
+function clonePanel(data) {
+	return {
+		type: CLONE_PANEL,
+		data,
+	};
+}
+
+export function addClonedPanel(data) {
+	return dispatch => dispatch(clonePanel(data));
 }
 
 function addPanelSet(data) {